requirements.txt
# requirements.txt
pandas
pyyaml
pytest